Renamed `test_dict.c` -> `test_dict.h`
